@tenderlove 2016-08-10
If we switch to using "buffer + `writev`" strategy in production, we might see better performance under load than just a simple benchmark.  The reason I say that is the allocator is probably under much greater pressure in production, especially as it keeps trying to resize the StringIO buffer before we write to the socket.