@arthurschreiber 2016-08-11
I did a bunch of benchmarking running against a live ernicorn instance, and the performance numbers I see are a bit disheartening. 😢

Here's my ernicorn handler:
<details>
```ruby
module Example
  def echo(*args)
    return *args
  end
end
```
</details>

And here's my benchmarking code:
<details>
```ruby
require "rubygems"
require "bundler/setup"

require 'benchmark'
require 'bertrpc'

ITER = 1_000

tiny = t[:ok, :awesome]
small = t[:ok, :answers, [42] * 42]
large = ["abc" * 1000] * 100
complex = [42, {:foo => 'bac' * 100}, t[(1..100).to_a]] * 10

svc = BERTRPC::Service.new('localhost', 9777)

Benchmark.bm do |bench|
  bench.report("BERT tiny") { ITER.times { svc.call.example.echo(tiny) } }
  bench.report("BERT small") { ITER.times { svc.call.example.echo(small) } }
  bench.report("BERT large") { ITER.times { svc.call.example.echo(large) } }
  bench.report("BERT complex") { ITER.times { svc.call.example.echo(complex) } }
end
```
</details>

Running both ernicorn and bertrpc with `BERT.encode`:

```
C02QW1XSG8WM:bertrpc arthurschreiber$ bundle exec ruby bench/bench.rb 
       user     system      total        real
BERT tiny  0.110000   0.080000   0.190000 (  0.460989)
BERT small  0.280000   0.080000   0.360000 (  0.809680)
BERT large  1.000000   0.540000   1.540000 (  6.549419)
BERT complex  4.670000   0.160000   4.830000 ( 10.299020)
```

Running both ernicorn and bertrpc with `BERT.encode_to_buffer`:

```
C02QW1XSG8WM:bertrpc arthurschreiber$ bundle exec ruby bench/bench.rb 
       user     system      total        real
BERT tiny  0.130000   0.180000   0.310000 (  0.602563)
BERT small  0.410000   0.600000   1.010000 (  1.722072)
BERT large  1.190000   1.760000   2.950000 (  8.664784)
BERT complex  6.780000  10.930000  17.710000 ( 28.507659)
```

Running both ernicorn and bertrpc with `BERT.encode_to_buffer` + `writev`:

```
C02QW1XSG8WM:bertrpc arthurschreiber$ bundle exec ruby bench/bench.rb 
       user     system      total        real
BERT tiny  0.120000   0.110000   0.230000 (  0.527234)
BERT small  0.340000   0.270000   0.610000 (  1.298251)
BERT large  1.000000   0.770000   1.770000 (  7.072049)
BERT complex  5.470000   3.900000   9.370000 ( 19.405413)
```

Running both ernicorn and bertrpc with `BERT.encode` and this patch:

```diff
diff --git a/lib/bert/encode.rb b/lib/bert/encode.rb
index 9302be6..db0c12e 100644
--- a/lib/bert/encode.rb
+++ b/lib/bert/encode.rb
@@ -62,6 +62,10 @@ module BERT
         @buf.each { |x| io.write x }
       end
 
+      def to_s
+        @buf.join("")
+      end
+
       def bytesize
         @buf.map(&:bytesize).inject :+
       end
@@ -74,11 +78,7 @@ module BERT
     end
 
     def self.encode(data)
-      buf = encode_to_buffer data
-      io = StringIO.new
-      io.set_encoding('binary') if io.respond_to?(:set_encoding)
-      buf.write_to io
-      io.string
+      encode_to_buffer(data).to_s
     end
 
     def self.encode_data(data, io)
```

```
C02QW1XSG8WM:bertrpc arthurschreiber$ bundle exec ruby bench/bench.rb 
       user     system      total        real
BERT tiny  0.100000   0.080000   0.180000 (  0.442146)
BERT small  0.280000   0.090000   0.370000 (  0.848133)
BERT large  0.880000   0.250000   1.130000 (  5.875011)
BERT complex  4.490000   0.160000   4.650000 ( 10.087896)
```

This `Buffer#to_s` patch also reduces memory usage. Not as much as using `BERT.encode_to_buffer` would (but it also does not impact performance negatively):

Before:
```
C02QW1XSG8WM:bert arthurschreiber$ bundle exec ruby bench/memsize.rb 
{:ORIGINAL=>53513430}
{:NEW=>4981285}
```

After:
```
C02QW1XSG8WM:bert arthurschreiber$ bundle exec ruby bench/memsize.rb 
{:ORIGINAL=>34231214}
{:NEW=>4981285}
```

Memory usage went down from `53513430` bytes to `34231214` bytes.