@carlosmn 2016-08-11
When we originally added `Buffer` I tried it out one of the repos that was giving us trouble in prod but it did not improve performance which is why the ernicorn is not using it. It turned out that the bottleneck was network mis-configuration together with inefficient code on the client so the code on the server/sender/encoding side was idle anyway. So maybe now it'll be better but I suspect the frontends might be more starved than the fileservers. E.g. the read replica for `github/github` is using 15GB out of 64GB of RAM and has a load average of 3-6, spread across 32 cores. API and fe hosts have half as much RAM and only 8 cores and are running a lower ratio of code that's ruby-but-actually-C.

The fundamental problem on the backend is still that we need to go through everything twice. That's probably what we're seeing on @arthurschreiber's `encode_to_buffer` benchmarks, especially on the complex structure with tiny objects. There it makes sense that sending everything into a buffer is quicker, since the pressure on the allocator would be small (especially if ruby keeps its own memory pool so it doesn't even involve syscalls). If you use larger payloads I would expect the buffer method to perform better. As @tenderlove points out, on prod it should perform better comparatively than on an idle laptop.

So then I wonder if writing to a temporary file, writing out the length header to the network socket and calling `sendfile` or `splice` on it would work out better. But this is also probably only worth it for larger payloads where we'd expect to have a bunch of syscalls anyway, and we don't generally have a lot of those. @brianmario is this in response to a particular performance problem or are you trying to be proactive about it?

Basically I think if we're hurting more on decoding on the client-side, where we're more constrained on resources and need to store the whole response from the server into memory before we'll give it to the decoder even though the format should be easy to handle in a serialized form.

Then there's things like the fact that we run charlock twice on blobs (at least on the API side) or charlock will spend two whole seconds on a 60MB buffer instead of deciding the page or two is probably good enough to know what the encoding is (maybe GitRPC should just be passing a smaller one there instead of expecting charlock to build up confidence and decide when it's OK to stop, I'm not sure).

Anyway, all this to say that if we're gonna go optimising syscalls, I think we can make more of an impact if we do it client-side, i.e. letting the client-side use a static buffer and make the decoder capable of acting on partial buffers.